pyproject.toml
[project]
name = "trinity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.1",
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "async-lru>=2.0.5",
    "ccxt>=4.5.33",
    "chromadb>=1.4.0",
    "duckdb>=1.4.3",
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "GitPython>=3.1.40",
    "google-api-core>=2.29.0",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.47.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-billing-budgets>=1.19.0",
    "google-cloud-dialogflow-cx>=2.1.0",
    "google-cloud-discoveryengine>=0.16.0",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "moviepy>=2.2.1",
    "numpy>=2.4.1",
    "oauthlib>=3.3.1",
    "pandas>=2.3.3",
    "polars>=1.37.1",
    "psutil>=7.2.1",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "rich>=14.2.0",
    "schedule>=1.2.2",
    "slowapi>=0.1.9",
    "tweepy>=4.16.0",
    "uvicorn[standard]>=0.40.0",
    "websockets>=15.0.1",
    "requests>=2.32.5",
    "mcp>=1.26.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.14.14",
]
